pyproject.toml
# SPDX-FileCopyrightText: 2024 Idiap Research Institute <contact@idiap.ch>
#
# SPDX-License-Identifier: BSD-3-Clause

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "arduino-pytwister"
version = "1.0.0"

requires-python = ">3.7"

dependencies = [
    "packaging",
    "pyserial",
]

authors = [
    {name = "Michael Liebling", email = "michael.liebling@idiap.ch"},
    {name = "François Marelli", email = "francois.marelli@idiap.ch"},
]
maintainers = [
    {name = "Daniel Carron", email = "daniel.carron@idiap.ch"}
]
description = "A python driver for arduino-controlled stepper motor"
readme = "README.md"
keywords = ["arduino", "sparkfun"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python"
]

[project.optional-dependencies]
qa = [
    "pre-commit",
    "reuse",
]

[project.urls]
Repository = ""

[tool.hatch.build.targets.wheel]
packages = ["src/pytwister"]
